change minimum Logger activity from DEBUG to NOTICE, fix using strings for Tags values

+++ src/InfluxDBJob.php
@@ -24,7 +24,7 @@ class InfluxDBJob
                 isset($event['value']) ? $event['value'] : 0,
                 isset($event['tags']) && is_array($event['tags']) ? $event['tags'] : [],
                 isset($event['fields']) && is_array($event['fields']) ? $event['fields'] : [],
-                isset($event['timestamp']) ? $event['timestamp'] : exec('date +%s%N')  // timestamp in nanoseconds on Linux ONLY
+                isset($event['timestamp']) ? (int)$event['timestamp'] : (int)exec('date +%s%N')  // timestamp in nanoseconds on Linux ONLY
             )
         ];
         try {
